+++ cmd/kahoy/config.go
@@ -56,6 +56,7 @@ type CmdConfig struct {
 		IncludeChanges           bool
 		ReportPath               string
 		AutoApprove              bool
+		CreateNamespace          bool
 	}
 }
 
@@ -93,6 +94,7 @@ func NewCmdConfig(args []string) (*CmdConfig, error) {
 	apply.Flag("include-changes", "Excludes all the resources without changes (old vs new states).").Short('f').BoolVar(&c.Apply.IncludeChanges)
 	apply.Flag("report-path", "Path to a file where the report data will be written, use `-` for stdout or nothing to disable").Short('r').StringVar(&c.Apply.ReportPath)
 	apply.Flag("auto-approve", "applies changes without asking for confirmation. Useful to run Kahoy on non interactive scenarios like CI.").BoolVar(&c.Apply.AutoApprove)
+	apply.Flag("create-namespace", "creates missing namespaces of the applied resources, used in regular and diff exacution modes.").BoolVar(&c.Apply.CreateNamespace)
 
 	// Deprecated flags.
 	apply.Flag("git-diff-filter", "DEPRECATED, use --include-changes.").Hidden().BoolVar(&c.Apply.IncludeChanges)