+++ CHANGELOG.md
@@ -13,9 +13,11 @@
 - Optional annotation filter for resources using Kubernetes standard label selectors using `--kube-include-annotation` flag.
 - Load `metav1.List` YAML resources as individual resources.
 - Allow groups waiting specific time after apply.
-- Add `fs-include` and `fs-exclude` arg options to kahoy app global configuration file as an alternative.
+- `fs-include` and `fs-exclude` arg options to kahoy app global configuration file as an alternative.
 - JSON report with the resources applied and deleted after the execution.
-- Confirmation prompt when running `kahoy apply` without diff or dry_run modes enabled. Optional flag `--auto-approve` to disable the prompt, useful for non interactive scenarios like CI.
+- Confirmation prompt when running `kahoy apply` without diff or dry-run modes enabled.
+- Optional `--auto-approve` to disable the confirmation prompt.
+- Optional `--create-namespace` on regular and diff modes that will create missing namespaces of applied resources.
 
 ### Changed